pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "evodrosophila"
version = "0.1.0"
description = "Evolved Connectome-Based Pattern Recognition from Drosophila Mushroom Body"
requires-python = ">=3.11"
dependencies = [
    "brian2>=2.5.4",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
